Move update of assertions after preprocessing
